refactor: remove `New` from refactored `ItemFormSection` component
