+++ components/quiz/ClozeQuestionModal.tsx
@@ -19,10 +19,10 @@ import { Dispatch, SetStateAction, useCallback, useMemo } from "react";
 import { PreloadedQuery } from "react-relay";
 import { useError } from "../ErrorContext";
 import { Form, FormSection } from "../Form";
-import ItemFormSectionNew, {
+import ItemFormSection, {
   CreateItem,
   Item,
-} from "../form-sections/item/ItemFormSectionNew";
+} from "../form-sections/item/ItemFormSection";
 import { FormErrors } from "../FormErrors";
 import { RichTextEditor, serializeToText } from "../RichTextEditor";
 import { EditRichTextButton } from "./EditRichTextButton";
@@ -150,7 +150,7 @@ export function ClozeQuestionModal({
       <DialogContent>
         <FormErrors error={error} />
         <Form>
-          <ItemFormSectionNew
+          <ItemFormSection
             operation="edit"
             item={item}
             setItem={setItem}
