+++ components/quiz/EditClozeQuestion.tsx
@@ -11,7 +11,7 @@ import {
   CreateItem,
   Item,
   mapRelayItemToItem,
-} from "../form-sections/item/ItemFormSectionNew";
+} from "../form-sections/item/ItemFormSection";
 import { ClozeQuestionData, ClozeQuestionModal } from "./ClozeQuestionModal";
 
 const ClozeQuestionFragment = graphql`
